# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BetterBet API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface User {
  id: number;
  displayName: string;
  createdAt: string;
}

export interface RegisterBody {
  displayName: string;
  /**
   * @minLength 4
   * @maxLength 4
   */
  pin: string;
}

export interface LoginBody {
  displayName: string;
  pin: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface Participant {
  id: number;
  userId: number;
  displayName: string;
  joinedAt: string;
}

export type ChallengeStatus =
  (typeof ChallengeStatus)[keyof typeof ChallengeStatus];

export const ChallengeStatus = {
  pending: "pending",
  active: "active",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Challenge {
  id: number;
  title: string;
  goal: string;
  wager: string;
  durationDays: number;
  requiredCheckinsPerWeek: number;
  createdBy: number;
  inviteCode: string;
  status: ChallengeStatus;
  startDate?: string | null;
  endDate?: string | null;
  createdAt: string;
}

export type ChallengeWithParticipants = Challenge & {
  participants: Participant[];
};

export interface CreateChallengeBody {
  title: string;
  goal: string;
  wager: string;
  durationDays: number;
  requiredCheckinsPerWeek: number;
}

export interface Checkin {
  id: number;
  challengeId: number;
  userId: number;
  displayName: string;
  date: string;
  photoUrl?: string | null;
  note?: string | null;
  createdAt: string;
}

export interface CreateCheckinBody {
  date: string;
  photoUrl?: string | null;
  note?: string | null;
}

export interface ParticipantProgress {
  userId: number;
  displayName: string;
  totalCheckins: number;
  requiredCheckins: number;
  currentWeekCheckins: number;
  requiredPerWeek: number;
  metGoal: boolean;
  streak: number;
  checkinDates: string[];
}

export interface ChallengeSummary {
  challenge: ChallengeWithParticipants;
  progress: ParticipantProgress[];
  daysRemaining: number;
  winner?: string | null;
  outcome?: string | null;
}

export interface UploadUrlBody {
  filename: string;
  contentType: string;
}

export interface UploadUrlResponse {
  uploadUrl: string;
  publicUrl: string;
}